Moved object-cache.php For Local Compatibility
When `object-cache.php` is by default at the root of the content directory it is loaded when in local installs as well. Unless the local host has memcache setup or at least the memcache extension installed on PHP it will cause fatal errors.

Created a `dropins` directory where `object-cache.php` and similar files can be stored. These files will automatically be moved into place by the `Boxfile` when deployed.